fix: layout ordering
