+++ web-registry/src/components/organisms/ProjectTopSection/ProjectTopSection.types.ts
@@ -1,5 +1,7 @@
 import { ProjectInfo } from '@regen-network/api/lib/generated/regen/ecocredit/v1/query';
 
+import { Party } from 'web-components/lib/components/modal/LedgerModal';
+
 import { Maybe, ProjectFieldsFragment } from 'generated/graphql';
 import { AllCreditClassQuery, SdgByIriQuery } from 'generated/sanity-graphql';
 import {
@@ -17,6 +19,9 @@ export type ProjectTopSectionProps = {
   geojson?: any;
   isGISFile?: boolean;
   onChainProjectId?: string;
+  landSteward?: Party;
+  landOwner?: Party;
+  projectDeveloper?: Party;
   loading?: boolean;
 };
